Reorganize entries. Refactor utils.
