+++ src/entries/map/reducer.js
